Modification cf bperr
